// sx-bi/src/app/analisis-de-ventas/models/ventaAcumulada.ts
export interface VentaAcumulada {
  fechaInicial: Date;
  fechaFinal: Date;
  clasificacion: Clasificacion;
  tipoVenta: string;
  tipo: string;
}

export enum Clasificacion {
  LINEA = 'LINEA',
  CLIENTE = 'CLIENTE',
  PRODUCTO = 'PRODUCTO',
  SUCURSAL = 'SUCURSAL',
  VENTA = 'VENTA',
  MES = 'MES'
}